+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- SPTrustedIdentityTokenIssuer
+  - Do not set property ProviderSignOutUri in SharePoint 2013 as it does
+  not exist
+
 ### Fixed
 
 - SPWebAppPolicy